+++ examples/padrino/app/app.rb
@@ -1,0 +1,16 @@
+module TestBh
+  class App < Padrino::Application
+    register Padrino::Helpers
+    register Padrino::Rendering
+
+    register Bh
+
+    get :root, map: '/' do
+      render :index
+    end
+
+    get :users, map: '/users' do
+      render :index
+    end
+  end
+end