+++ lib/bh.rb
@@ -7,5 +7,7 @@ helpers_folder = File.expand_path '../bh/helpers/*_helper.rb', __FILE__
 Dir[helpers_folder].each{|file| require file}
 
 # Conditionally require platform-specific helpers
-require 'bh/core_ext/railtie' if defined?(Rails)
-require 'bh/core_ext/middleman' if defined?(Middleman)
+extension   = :railtie if defined?(Rails)
+extension ||= :middleman if defined?(Middleman)
+extension ||= :padrino if defined?(Padrino)
+require "bh/core_ext/#{extension}" if extension